+++ CHANGELOG.md
@@ -6,6 +6,10 @@
     - `seqkit faidx`: allow empty lines at the end of sequences.
     - `seqkit stats`: fix compute Q1 and Q3 of sequence length for one record. [#353](https://github.com/shenwei356/seqkit/issues/353)
     - `seqkit seq`: allow filtering sequences of length zero. thanks to @penglbio.
+    - `seqkit rename`:
+        - new flag `-s/--separator` for setting separater between original ID/name and the counter. [#360](https://github.com/shenwei356/seqkit/issues/360)
+        - new flag `-N/--start-num` for starting count number for duplicated IDs/names. [#360](https://github.com/shenwei356/seqkit/issues/360)
+        - do not append space if there's no description after the sequene ID.
 - [SeqKit v2.3.1](https://github.com/shenwei356/seqkit/releases/tag/v2.3.1) - 2022-09-22
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v2.3.1/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v2.3.1)
     - `seqkit grep/locate`: fix bug of FMIndex building for empty sequences. [#321](https://github.com/shenwei356/seqkit/issues/321)